Update cmdable to be compatible with v6.14
