+++ cmdable.go
@@ -10,9 +10,7 @@ func (m *ClientMock) Pipeline() redis.Pipeliner {
 		return m.client.Pipeline()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(redis.Pipeliner)
+	return m.Called().Get(0).(redis.Pipeliner)
 }
 
 func (m *ClientMock) Pipelined(fn func(redis.Pipeliner) error) ([]redis.Cmder, error) {
@@ -40,9 +38,7 @@ func (m *ClientMock) TxPipeline() redis.Pipeliner {
 		return m.client.TxPipeline()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(redis.Pipeliner)
+	return m.Called().Get(0).(redis.Pipeliner)
 }
 
 func (m *ClientMock) ClientGetName() *redis.StringCmd {
@@ -50,9 +46,7 @@ func (m *ClientMock) ClientGetName() *redis.StringCmd {
 		return m.client.ClientGetName()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Echo(message interface{}) *redis.StringCmd {
@@ -60,9 +54,7 @@ func (m *ClientMock) Echo(message interface{}) *redis.StringCmd {
 		return m.client.Echo(message)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Ping() *redis.StatusCmd {
@@ -70,9 +62,7 @@ func (m *ClientMock) Ping() *redis.StatusCmd {
 		return m.client.Ping()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Quit() *redis.StatusCmd {
@@ -80,9 +70,7 @@ func (m *ClientMock) Quit() *redis.StatusCmd {
 		return m.client.Quit()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Del(keys ...string) *redis.IntCmd {
@@ -90,9 +78,7 @@ func (m *ClientMock) Del(keys ...string) *redis.IntCmd {
 		return m.client.Del(keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Unlink(keys ...string) *redis.IntCmd {
@@ -100,9 +86,7 @@ func (m *ClientMock) Unlink(keys ...string) *redis.IntCmd {
 		return m.client.Unlink(keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Dump(key string) *redis.StringCmd {
@@ -110,9 +94,7 @@ func (m *ClientMock) Dump(key string) *redis.StringCmd {
 		return m.client.Dump(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Exists(keys ...string) *redis.IntCmd {
@@ -120,9 +102,7 @@ func (m *ClientMock) Exists(keys ...string) *redis.IntCmd {
 		return m.client.Exists(keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Expire(key string, expiration time.Duration) *redis.BoolCmd {
@@ -130,9 +110,7 @@ func (m *ClientMock) Expire(key string, expiration time.Duration) *redis.BoolCmd
 		return m.client.Expire(key, expiration)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) ExpireAt(key string, tm time.Time) *redis.BoolCmd {
@@ -140,9 +118,7 @@ func (m *ClientMock) ExpireAt(key string, tm time.Time) *redis.BoolCmd {
 		return m.client.ExpireAt(key, tm)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) Keys(pattern string) *redis.StringSliceCmd {
@@ -150,9 +126,7 @@ func (m *ClientMock) Keys(pattern string) *redis.StringSliceCmd {
 		return m.client.Keys(pattern)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) Migrate(host, port, key string, db int64, timeout time.Duration) *redis.StatusCmd {
@@ -160,9 +134,7 @@ func (m *ClientMock) Migrate(host, port, key string, db int64, timeout time.Dura
 		return m.client.Migrate(host, port, key, db, timeout)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Move(key string, db int64) *redis.BoolCmd {
@@ -170,9 +142,7 @@ func (m *ClientMock) Move(key string, db int64) *redis.BoolCmd {
 		return m.client.Move(key, db)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) ObjectRefCount(key string) *redis.IntCmd {
@@ -180,9 +150,7 @@ func (m *ClientMock) ObjectRefCount(key string) *redis.IntCmd {
 		return m.client.ObjectRefCount(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ObjectEncoding(key string) *redis.StringCmd {
@@ -190,9 +158,7 @@ func (m *ClientMock) ObjectEncoding(key string) *redis.StringCmd {
 		return m.client.ObjectEncoding(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) ObjectIdleTime(key string) *redis.DurationCmd {
@@ -200,9 +166,7 @@ func (m *ClientMock) ObjectIdleTime(key string) *redis.DurationCmd {
 		return m.client.ObjectIdleTime(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.DurationCmd)
+	return m.Called().Get(0).(*redis.DurationCmd)
 }
 
 func (m *ClientMock) Persist(key string) *redis.BoolCmd {
@@ -210,9 +174,7 @@ func (m *ClientMock) Persist(key string) *redis.BoolCmd {
 		return m.client.Persist(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) PExpire(key string, expiration time.Duration) *redis.BoolCmd {
@@ -220,9 +182,7 @@ func (m *ClientMock) PExpire(key string, expiration time.Duration) *redis.BoolCm
 		return m.client.PExpire(key, expiration)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) PExpireAt(key string, tm time.Time) *redis.BoolCmd {
@@ -230,9 +190,7 @@ func (m *ClientMock) PExpireAt(key string, tm time.Time) *redis.BoolCmd {
 		return m.client.PExpireAt(key, tm)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) PTTL(key string) *redis.DurationCmd {
@@ -240,9 +198,7 @@ func (m *ClientMock) PTTL(key string) *redis.DurationCmd {
 		return m.client.PTTL(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.DurationCmd)
+	return m.Called().Get(0).(*redis.DurationCmd)
 }
 
 func (m *ClientMock) RandomKey() *redis.StringCmd {
@@ -250,9 +206,7 @@ func (m *ClientMock) RandomKey() *redis.StringCmd {
 		return m.client.RandomKey()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Rename(key, newkey string) *redis.StatusCmd {
@@ -260,9 +214,7 @@ func (m *ClientMock) Rename(key, newkey string) *redis.StatusCmd {
 		return m.client.Rename(key, newkey)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) RenameNX(key, newkey string) *redis.BoolCmd {
@@ -270,9 +222,7 @@ func (m *ClientMock) RenameNX(key, newkey string) *redis.BoolCmd {
 		return m.client.RenameNX(key, newkey)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) Restore(key string, ttl time.Duration, value string) *redis.StatusCmd {
@@ -280,9 +230,7 @@ func (m *ClientMock) Restore(key string, ttl time.Duration, value string) *redis
 		return m.client.Restore(key, ttl, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) RestoreReplace(key string, ttl time.Duration, value string) *redis.StatusCmd {
@@ -290,9 +238,7 @@ func (m *ClientMock) RestoreReplace(key string, ttl time.Duration, value string)
 		return m.client.RestoreReplace(key, ttl, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Sort(key string, sort *redis.Sort) *redis.StringSliceCmd {
@@ -300,9 +246,7 @@ func (m *ClientMock) Sort(key string, sort *redis.Sort) *redis.StringSliceCmd {
 		return m.client.Sort(key, sort)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) SortStore(key, store string, sort *redis.Sort) *redis.IntCmd {
@@ -310,9 +254,7 @@ func (m *ClientMock) SortStore(key, store string, sort *redis.Sort) *redis.IntCm
 		return m.client.SortStore(key, store, sort)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) SortInterfaces(key string, sort *redis.Sort) *redis.SliceCmd {
@@ -320,9 +262,7 @@ func (m *ClientMock) SortInterfaces(key string, sort *redis.Sort) *redis.SliceCm
 		return m.client.SortInterfaces(key, sort)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.SliceCmd)
+	return m.Called().Get(0).(*redis.SliceCmd)
 }
 
 func (m *ClientMock) Touch(keys ...string) *redis.IntCmd {
@@ -330,9 +270,7 @@ func (m *ClientMock) Touch(keys ...string) *redis.IntCmd {
 		return m.client.Touch()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) TTL(key string) *redis.DurationCmd {
@@ -340,9 +278,7 @@ func (m *ClientMock) TTL(key string) *redis.DurationCmd {
 		return m.client.TTL(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.DurationCmd)
+	return m.Called().Get(0).(*redis.DurationCmd)
 }
 
 func (m *ClientMock) Type(key string) *redis.StatusCmd {
@@ -350,9 +286,7 @@ func (m *ClientMock) Type(key string) *redis.StatusCmd {
 		return m.client.Type(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Scan(cursor uint64, match string, count int64) *redis.ScanCmd {
@@ -360,9 +294,7 @@ func (m *ClientMock) Scan(cursor uint64, match string, count int64) *redis.ScanC
 		return m.client.Scan(cursor, match, count)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ScanCmd)
+	return m.Called().Get(0).(*redis.ScanCmd)
 }
 
 func (m *ClientMock) SScan(key string, cursor uint64, match string, count int64) *redis.ScanCmd {
@@ -370,29 +302,7 @@ func (m *ClientMock) SScan(key string, cursor uint64, match string, count int64)
 		return m.client.SScan(key, cursor, match, count)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ScanCmd)
-}
-
-func (m *ClientMock) HScan(key string, cursor uint64, match string, count int64) *redis.ScanCmd {
-	if !m.hasStub("HScan") {
-		return m.client.HScan(key, cursor, match, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ScanCmd)
-}
-
-func (m *ClientMock) ZScan(key string, cursor uint64, match string, count int64) *redis.ScanCmd {
-	if !m.hasStub("ZScan") {
-		return m.client.ZScan(key, cursor, match, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ScanCmd)
+	return m.Called().Get(0).(*redis.ScanCmd)
 }
 
 func (m *ClientMock) Append(key, value string) *redis.IntCmd {
@@ -400,9 +310,7 @@ func (m *ClientMock) Append(key, value string) *redis.IntCmd {
 		return m.client.Append(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitCount(key string, bitCount *redis.BitCount) *redis.IntCmd {
@@ -410,9 +318,7 @@ func (m *ClientMock) BitCount(key string, bitCount *redis.BitCount) *redis.IntCm
 		return m.client.BitCount(key, bitCount)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitOpAnd(destKey string, keys ...string) *redis.IntCmd {
@@ -420,9 +326,7 @@ func (m *ClientMock) BitOpAnd(destKey string, keys ...string) *redis.IntCmd {
 		return m.client.BitOpAnd(destKey, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitOpOr(destKey string, keys ...string) *redis.IntCmd {
@@ -430,9 +334,7 @@ func (m *ClientMock) BitOpOr(destKey string, keys ...string) *redis.IntCmd {
 		return m.client.BitOpOr(destKey, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitOpXor(destKey string, keys ...string) *redis.IntCmd {
@@ -440,9 +342,7 @@ func (m *ClientMock) BitOpXor(destKey string, keys ...string) *redis.IntCmd {
 		return m.client.BitOpXor(destKey, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitOpNot(destKey string, key string) *redis.IntCmd {
@@ -450,9 +350,7 @@ func (m *ClientMock) BitOpNot(destKey string, key string) *redis.IntCmd {
 		return m.client.BitOpNot(destKey, key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BitPos(key string, bit int64, pos ...int64) *redis.IntCmd {
@@ -460,9 +358,7 @@ func (m *ClientMock) BitPos(key string, bit int64, pos ...int64) *redis.IntCmd {
 		return m.client.BitPos(key, bit, pos...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Decr(key string) *redis.IntCmd {
@@ -470,9 +366,7 @@ func (m *ClientMock) Decr(key string) *redis.IntCmd {
 		return m.client.Decr(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) DecrBy(key string, decrement int64) *redis.IntCmd {
@@ -480,9 +374,7 @@ func (m *ClientMock) DecrBy(key string, decrement int64) *redis.IntCmd {
 		return m.client.DecrBy(key, decrement)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Get(key string) *redis.StringCmd {
@@ -490,9 +382,7 @@ func (m *ClientMock) Get(key string) *redis.StringCmd {
 		return m.client.Get(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) GetBit(key string, offset int64) *redis.IntCmd {
@@ -500,9 +390,7 @@ func (m *ClientMock) GetBit(key string, offset int64) *redis.IntCmd {
 		return m.client.GetBit(key, offset)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) GetRange(key string, start, end int64) *redis.StringCmd {
@@ -510,9 +398,7 @@ func (m *ClientMock) GetRange(key string, start, end int64) *redis.StringCmd {
 		return m.client.GetRange(key, start, end)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) GetSet(key string, value interface{}) *redis.StringCmd {
@@ -520,9 +406,7 @@ func (m *ClientMock) GetSet(key string, value interface{}) *redis.StringCmd {
 		return m.client.GetSet(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Incr(key string) *redis.IntCmd {
@@ -530,9 +414,7 @@ func (m *ClientMock) Incr(key string) *redis.IntCmd {
 		return m.client.Incr(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) IncrBy(key string, value int64) *redis.IntCmd {
@@ -540,9 +422,7 @@ func (m *ClientMock) IncrBy(key string, value int64) *redis.IntCmd {
 		return m.client.IncrBy(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) IncrByFloat(key string, value float64) *redis.FloatCmd {
@@ -550,9 +430,7 @@ func (m *ClientMock) IncrByFloat(key string, value float64) *redis.FloatCmd {
 		return m.client.IncrByFloat(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
+	return m.Called().Get(0).(*redis.FloatCmd)
 }
 
 func (m *ClientMock) MGet(keys ...string) *redis.SliceCmd {
@@ -560,9 +438,7 @@ func (m *ClientMock) MGet(keys ...string) *redis.SliceCmd {
 		return m.client.MGet()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.SliceCmd)
+	return m.Called().Get(0).(*redis.SliceCmd)
 }
 
 func (m *ClientMock) MSet(pairs ...interface{}) *redis.StatusCmd {
@@ -570,9 +446,7 @@ func (m *ClientMock) MSet(pairs ...interface{}) *redis.StatusCmd {
 		return m.client.MSet()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) MSetNX(pairs ...interface{}) *redis.BoolCmd {
@@ -580,9 +454,7 @@ func (m *ClientMock) MSetNX(pairs ...interface{}) *redis.BoolCmd {
 		return m.client.MSetNX()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) Set(key string, value interface{}, expiration time.Duration) *redis.StatusCmd {
@@ -590,9 +462,7 @@ func (m *ClientMock) Set(key string, value interface{}, expiration time.Duration
 		return m.client.Set(key, value, expiration)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) SetBit(key string, offset int64, value int) *redis.IntCmd {
@@ -600,9 +470,7 @@ func (m *ClientMock) SetBit(key string, offset int64, value int) *redis.IntCmd {
 		return m.client.SetBit(key, offset, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) SetNX(key string, value interface{}, expiration time.Duration) *redis.BoolCmd {
@@ -610,9 +478,7 @@ func (m *ClientMock) SetNX(key string, value interface{}, expiration time.Durati
 		return m.client.SetNX(key, value, expiration)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) SetXX(key string, value interface{}, expiration time.Duration) *redis.BoolCmd {
@@ -620,9 +486,7 @@ func (m *ClientMock) SetXX(key string, value interface{}, expiration time.Durati
 		return m.client.SetXX(key, value, expiration)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) SetRange(key string, offset int64, value string) *redis.IntCmd {
@@ -630,9 +494,7 @@ func (m *ClientMock) SetRange(key string, offset int64, value string) *redis.Int
 		return m.client.SetRange(key, offset, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) StrLen(key string) *redis.IntCmd {
@@ -640,139 +502,7 @@ func (m *ClientMock) StrLen(key string) *redis.IntCmd {
 		return m.client.StrLen(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) HDel(key string, fields ...string) *redis.IntCmd {
-	if !m.hasStub("HDel") {
-		return m.client.HDel(key, fields...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) HExists(key, field string) *redis.BoolCmd {
-	if !m.hasStub("HExists") {
-		return m.client.HExists(key, field)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
-}
-
-func (m *ClientMock) HGet(key, field string) *redis.StringCmd {
-	if !m.hasStub("HGet") {
-		return m.client.HGet(key, field)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
-}
-
-func (m *ClientMock) HGetAll(key string) *redis.StringStringMapCmd {
-	if !m.hasStub("HGetAll") {
-		return m.client.HGetAll(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringStringMapCmd)
-}
-
-func (m *ClientMock) HIncrBy(key, field string, incr int64) *redis.IntCmd {
-	if !m.hasStub("HIncrBy") {
-		return m.client.HIncrBy(key, field, incr)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) HIncrByFloat(key, field string, incr float64) *redis.FloatCmd {
-	if !m.hasStub("HIncrByFloat") {
-		return m.client.HIncrByFloat(key, field, incr)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) HKeys(key string) *redis.StringSliceCmd {
-	if !m.hasStub("HKeys") {
-		return m.client.HKeys(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) HLen(key string) *redis.IntCmd {
-	if !m.hasStub("HLen") {
-		return m.client.HLen(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) HMGet(key string, fields ...string) *redis.SliceCmd {
-	if !m.hasStub("HMGet") {
-		return m.client.HMGet(key, fields...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.SliceCmd)
-}
-
-func (m *ClientMock) HMSet(key string, fields map[string]interface{}) *redis.StatusCmd {
-	if !m.hasStub("HMSet") {
-		return m.client.HMSet(key, fields)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
-}
-
-func (m *ClientMock) HSet(key, field string, value interface{}) *redis.BoolCmd {
-	if !m.hasStub("HSet") {
-		return m.client.HSet(key, field, value)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
-}
-
-func (m *ClientMock) HSetNX(key, field string, value interface{}) *redis.BoolCmd {
-	if !m.hasStub("HSetNX") {
-		return m.client.HSetNX(key, field, value)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
-}
-
-func (m *ClientMock) HVals(key string) *redis.StringSliceCmd {
-	if !m.hasStub("HVals") {
-		return m.client.HVals(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) BLPop(timeout time.Duration, keys ...string) *redis.StringSliceCmd {
@@ -780,9 +510,7 @@ func (m *ClientMock) BLPop(timeout time.Duration, keys ...string) *redis.StringS
 		return m.client.BLPop(timeout, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) BRPop(timeout time.Duration, keys ...string) *redis.StringSliceCmd {
@@ -790,9 +518,7 @@ func (m *ClientMock) BRPop(timeout time.Duration, keys ...string) *redis.StringS
 		return m.client.BRPop(timeout, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) BRPopLPush(source, destination string, timeout time.Duration) *redis.StringCmd {
@@ -800,9 +526,7 @@ func (m *ClientMock) BRPopLPush(source, destination string, timeout time.Duratio
 		return m.client.BRPopLPush(source, destination, timeout)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) LIndex(key string, index int64) *redis.StringCmd {
@@ -810,9 +534,7 @@ func (m *ClientMock) LIndex(key string, index int64) *redis.StringCmd {
 		return m.client.LIndex(key, index)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) LInsert(key, op string, pivot, value interface{}) *redis.IntCmd {
@@ -820,9 +542,7 @@ func (m *ClientMock) LInsert(key, op string, pivot, value interface{}) *redis.In
 		return m.client.LInsert(key, op, pivot, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LInsertBefore(key string, pivot, value interface{}) *redis.IntCmd {
@@ -830,9 +550,7 @@ func (m *ClientMock) LInsertBefore(key string, pivot, value interface{}) *redis.
 		return m.client.LInsertBefore(key, pivot, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LInsertAfter(key string, pivot, value interface{}) *redis.IntCmd {
@@ -840,9 +558,7 @@ func (m *ClientMock) LInsertAfter(key string, pivot, value interface{}) *redis.I
 		return m.client.LInsertAfter(key, pivot, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LLen(key string) *redis.IntCmd {
@@ -850,9 +566,7 @@ func (m *ClientMock) LLen(key string) *redis.IntCmd {
 		return m.client.LLen(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LPop(key string) *redis.StringCmd {
@@ -860,9 +574,7 @@ func (m *ClientMock) LPop(key string) *redis.StringCmd {
 		return m.client.LPop(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) LPush(key string, values ...interface{}) *redis.IntCmd {
@@ -870,9 +582,7 @@ func (m *ClientMock) LPush(key string, values ...interface{}) *redis.IntCmd {
 		return m.client.LPush(key, values...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LPushX(key string, value interface{}) *redis.IntCmd {
@@ -880,9 +590,7 @@ func (m *ClientMock) LPushX(key string, value interface{}) *redis.IntCmd {
 		return m.client.LPushX(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LRange(key string, start, stop int64) *redis.StringSliceCmd {
@@ -890,9 +598,7 @@ func (m *ClientMock) LRange(key string, start, stop int64) *redis.StringSliceCmd
 		return m.client.LRange(key, start, stop)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) LRem(key string, count int64, value interface{}) *redis.IntCmd {
@@ -900,9 +606,7 @@ func (m *ClientMock) LRem(key string, count int64, value interface{}) *redis.Int
 		return m.client.LRem(key, count, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) LSet(key string, index int64, value interface{}) *redis.StatusCmd {
@@ -910,9 +614,7 @@ func (m *ClientMock) LSet(key string, index int64, value interface{}) *redis.Sta
 		return m.client.LSet(key, index, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) LTrim(key string, start, stop int64) *redis.StatusCmd {
@@ -920,9 +622,7 @@ func (m *ClientMock) LTrim(key string, start, stop int64) *redis.StatusCmd {
 		return m.client.LTrim(key, start, stop)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) RPop(key string) *redis.StringCmd {
@@ -930,9 +630,7 @@ func (m *ClientMock) RPop(key string) *redis.StringCmd {
 		return m.client.RPop(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) RPopLPush(source, destination string) *redis.StringCmd {
@@ -940,9 +638,7 @@ func (m *ClientMock) RPopLPush(source, destination string) *redis.StringCmd {
 		return m.client.RPopLPush(source, destination)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) RPush(key string, values ...interface{}) *redis.IntCmd {
@@ -950,9 +646,7 @@ func (m *ClientMock) RPush(key string, values ...interface{}) *redis.IntCmd {
 		return m.client.RPush(key, values...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) RPushX(key string, value interface{}) *redis.IntCmd {
@@ -960,499 +654,7 @@ func (m *ClientMock) RPushX(key string, value interface{}) *redis.IntCmd {
 		return m.client.RPushX(key, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SAdd(key string, members ...interface{}) *redis.IntCmd {
-	if !m.hasStub("SAdd") {
-		return m.client.SAdd(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SCard(key string) *redis.IntCmd {
-	if !m.hasStub("SCard") {
-		return m.client.SCard(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SDiff(keys ...string) *redis.StringSliceCmd {
-	if !m.hasStub("SDiff") {
-		return m.client.SDiff()
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SDiffStore(destination string, keys ...string) *redis.IntCmd {
-	if !m.hasStub("SDiffStore") {
-		return m.client.SDiffStore(destination, keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SInter(keys ...string) *redis.StringSliceCmd {
-	if !m.hasStub("SInter") {
-		return m.client.SInter()
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SInterStore(destination string, keys ...string) *redis.IntCmd {
-	if !m.hasStub("SInterStore") {
-		return m.client.SInterStore(destination, keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SIsMember(key string, member interface{}) *redis.BoolCmd {
-	if !m.hasStub("SIsMember") {
-		return m.client.SIsMember(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
-}
-
-func (m *ClientMock) SMembers(key string) *redis.StringSliceCmd {
-	if !m.hasStub("SMembers") {
-		return m.client.SMembers(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SMembersMap(key string) *redis.StringStructMapCmd {
-	if !m.hasStub("SMembersMap") {
-		return m.client.SMembersMap(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringStructMapCmd)
-}
-
-func (m *ClientMock) SMove(source, destination string, member interface{}) *redis.BoolCmd {
-	if !m.hasStub("SMove") {
-		return m.client.SMove(source, destination, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
-}
-
-func (m *ClientMock) SPop(key string) *redis.StringCmd {
-	if !m.hasStub("SPop") {
-		return m.client.SPop(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
-}
-
-func (m *ClientMock) SPopN(key string, count int64) *redis.StringSliceCmd {
-	if !m.hasStub("SPopN") {
-		return m.client.SPopN(key, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SRandMember(key string) *redis.StringCmd {
-	if !m.hasStub("SRandMember") {
-		return m.client.SRandMember(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
-}
-
-func (m *ClientMock) SRandMemberN(key string, count int64) *redis.StringSliceCmd {
-	if !m.hasStub("SRandMemberN") {
-		return m.client.SRandMemberN(key, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SRem(key string, members ...interface{}) *redis.IntCmd {
-	if !m.hasStub("SRem") {
-		return m.client.SRem(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) SUnion(keys ...string) *redis.StringSliceCmd {
-	if !m.hasStub("SUnion") {
-		return m.client.SUnion(keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) SUnionStore(destination string, keys ...string) *redis.IntCmd {
-	if !m.hasStub("SUnionStore") {
-		return m.client.SUnionStore(destination, keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAdd(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAdd") {
-		return m.client.ZAdd(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAddNX(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAddNX") {
-		return m.client.ZAddNX(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAddXX(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAddXX") {
-		return m.client.ZAddXX(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAddCh(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAddCh") {
-		return m.client.ZAddCh(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAddNXCh(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAddNXCh") {
-		return m.client.ZAddNXCh(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZAddXXCh(key string, members ...redis.Z) *redis.IntCmd {
-	if !m.hasStub("ZAddXXCh") {
-		return m.client.ZAddXXCh(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZIncr(key string, member redis.Z) *redis.FloatCmd {
-	if !m.hasStub("ZIncr") {
-		return m.client.ZIncr(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) ZIncrNX(key string, member redis.Z) *redis.FloatCmd {
-	if !m.hasStub("ZIncrNX") {
-		return m.client.ZIncrNX(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) ZIncrXX(key string, member redis.Z) *redis.FloatCmd {
-	if !m.hasStub("ZIncrXX") {
-		return m.client.ZIncrXX(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) ZCard(key string) *redis.IntCmd {
-	if !m.hasStub("ZCard") {
-		return m.client.ZCard(key)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZCount(key, min, max string) *redis.IntCmd {
-	if !m.hasStub("ZCount") {
-		return m.client.ZCount(key, min, max)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZLexCount(key, min, max string) *redis.IntCmd {
-	if !m.hasStub("ZLexCount") {
-		return m.client.ZLexCount(key, min, max)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZIncrBy(key string, increment float64, member string) *redis.FloatCmd {
-	if !m.hasStub("ZIncrBy") {
-		return m.client.ZIncrBy(key, increment, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) ZInterStore(destination string, store redis.ZStore, keys ...string) *redis.IntCmd {
-	if !m.hasStub("ZInterStore") {
-		return m.client.ZInterStore(destination, store, keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRange(key string, start, stop int64) *redis.StringSliceCmd {
-	if !m.hasStub("ZRange") {
-		return m.client.ZRange(key, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRangeWithScores(key string, start, stop int64) *redis.ZSliceCmd {
-	if !m.hasStub("ZRangeWithScores") {
-		return m.client.ZRangeWithScores(key, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ZSliceCmd)
-}
-
-func (m *ClientMock) ZRangeByScore(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
-	if !m.hasStub("ZRangeByScore") {
-		return m.client.ZRangeByScore(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRangeByLex(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
-	if !m.hasStub("ZRangeByLex") {
-		return m.client.ZRangeByLex(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRangeByScoreWithScores(key string, opt redis.ZRangeBy) *redis.ZSliceCmd {
-	if !m.hasStub("ZRangeByScoreWithScores") {
-		return m.client.ZRangeByScoreWithScores(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ZSliceCmd)
-}
-
-func (m *ClientMock) ZRank(key, member string) *redis.IntCmd {
-	if !m.hasStub("ZRank") {
-		return m.client.ZRank(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRem(key string, members ...interface{}) *redis.IntCmd {
-	if !m.hasStub("ZRem") {
-		return m.client.ZRem(key, members...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRemRangeByRank(key string, start, stop int64) *redis.IntCmd {
-	if !m.hasStub("ZRemRangeByRank") {
-		return m.client.ZRemRangeByRank(key, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRemRangeByScore(key, min, max string) *redis.IntCmd {
-	if !m.hasStub("ZRemRangeByScore") {
-		return m.client.ZRemRangeByScore(key, min, max)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRemRangeByLex(key, min, max string) *redis.IntCmd {
-	if !m.hasStub("ZRemRangeByLex") {
-		return m.client.ZRemRangeByLex(key, min, max)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZRevRange(key string, start, stop int64) *redis.StringSliceCmd {
-	if !m.hasStub("ZRevRange") {
-		return m.client.ZRevRange(key, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRevRangeWithScores(key string, start, stop int64) *redis.ZSliceCmd {
-	if !m.hasStub("ZRevRangeWithScores") {
-		return m.client.ZRevRangeWithScores(key, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ZSliceCmd)
-}
-
-func (m *ClientMock) ZRevRangeByScore(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
-	if !m.hasStub("ZRevRangeByScore") {
-		return m.client.ZRevRangeByScore(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRevRangeByLex(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
-	if !m.hasStub("ZRevRangeByLex") {
-		return m.client.ZRevRangeByLex(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
-}
-
-func (m *ClientMock) ZRevRangeByScoreWithScores(key string, opt redis.ZRangeBy) *redis.ZSliceCmd {
-	if !m.hasStub("ZRevRangeByScoreWithScores") {
-		return m.client.ZRevRangeByScoreWithScores(key, opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ZSliceCmd)
-}
-
-func (m *ClientMock) ZRevRank(key, member string) *redis.IntCmd {
-	if !m.hasStub("ZRevRank") {
-		return m.client.ZRevRank(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) ZScore(key, member string) *redis.FloatCmd {
-	if !m.hasStub("ZScore") {
-		return m.client.ZScore(key, member)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
-}
-
-func (m *ClientMock) ZUnionStore(dest string, store redis.ZStore, keys ...string) *redis.IntCmd {
-	if !m.hasStub("ZUnionStore") {
-		return m.client.ZUnionStore(dest, store, keys...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) PFAdd(key string, els ...interface{}) *redis.IntCmd {
@@ -1460,9 +662,7 @@ func (m *ClientMock) PFAdd(key string, els ...interface{}) *redis.IntCmd {
 		return m.client.PFAdd(key, els...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) PFCount(keys ...string) *redis.IntCmd {
@@ -1470,9 +670,7 @@ func (m *ClientMock) PFCount(keys ...string) *redis.IntCmd {
 		return m.client.PFCount()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) PFMerge(dest string, keys ...string) *redis.StatusCmd {
@@ -1480,9 +678,7 @@ func (m *ClientMock) PFMerge(dest string, keys ...string) *redis.StatusCmd {
 		return m.client.PFMerge(dest, keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) BgRewriteAOF() *redis.StatusCmd {
@@ -1490,9 +686,7 @@ func (m *ClientMock) BgRewriteAOF() *redis.StatusCmd {
 		return m.client.BgRewriteAOF()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) BgSave() *redis.StatusCmd {
@@ -1500,9 +694,7 @@ func (m *ClientMock) BgSave() *redis.StatusCmd {
 		return m.client.BgSave()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClientKill(ipPort string) *redis.StatusCmd {
@@ -1510,9 +702,7 @@ func (m *ClientMock) ClientKill(ipPort string) *redis.StatusCmd {
 		return m.client.ClientKill(ipPort)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClientList() *redis.StringCmd {
@@ -1520,9 +710,7 @@ func (m *ClientMock) ClientList() *redis.StringCmd {
 		return m.client.ClientList()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) ClientPause(dur time.Duration) *redis.BoolCmd {
@@ -1530,9 +718,7 @@ func (m *ClientMock) ClientPause(dur time.Duration) *redis.BoolCmd {
 		return m.client.ClientPause(dur)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolCmd)
+	return m.Called().Get(0).(*redis.BoolCmd)
 }
 
 func (m *ClientMock) ConfigGet(parameter string) *redis.SliceCmd {
@@ -1540,9 +726,7 @@ func (m *ClientMock) ConfigGet(parameter string) *redis.SliceCmd {
 		return m.client.ConfigGet(parameter)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.SliceCmd)
+	return m.Called().Get(0).(*redis.SliceCmd)
 }
 
 func (m *ClientMock) ConfigResetStat() *redis.StatusCmd {
@@ -1550,9 +734,7 @@ func (m *ClientMock) ConfigResetStat() *redis.StatusCmd {
 		return m.client.ConfigResetStat()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ConfigSet(parameter, value string) *redis.StatusCmd {
@@ -1560,9 +742,7 @@ func (m *ClientMock) ConfigSet(parameter, value string) *redis.StatusCmd {
 		return m.client.ConfigSet(parameter, value)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ConfigRewrite() *redis.StatusCmd {
@@ -1570,9 +750,7 @@ func (m *ClientMock) ConfigRewrite() *redis.StatusCmd {
 		return m.client.ConfigRewrite()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) DBSize() *redis.IntCmd {
@@ -1580,9 +758,7 @@ func (m *ClientMock) DBSize() *redis.IntCmd {
 		return m.client.DBSize()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) FlushAll() *redis.StatusCmd {
@@ -1590,9 +766,7 @@ func (m *ClientMock) FlushAll() *redis.StatusCmd {
 		return m.client.FlushAll()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) FlushAllAsync() *redis.StatusCmd {
@@ -1600,9 +774,7 @@ func (m *ClientMock) FlushAllAsync() *redis.StatusCmd {
 		return m.client.FlushAllAsync()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) FlushDB() *redis.StatusCmd {
@@ -1610,9 +782,7 @@ func (m *ClientMock) FlushDB() *redis.StatusCmd {
 		return m.client.FlushDB()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) FlushDBAsync() *redis.StatusCmd {
@@ -1620,9 +790,7 @@ func (m *ClientMock) FlushDBAsync() *redis.StatusCmd {
 		return m.client.FlushDBAsync()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Info(section ...string) *redis.StringCmd {
@@ -1630,9 +798,7 @@ func (m *ClientMock) Info(section ...string) *redis.StringCmd {
 		return m.client.Info()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) LastSave() *redis.IntCmd {
@@ -1640,9 +806,7 @@ func (m *ClientMock) LastSave() *redis.IntCmd {
 		return m.client.LastSave()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) Save() *redis.StatusCmd {
@@ -1650,9 +814,7 @@ func (m *ClientMock) Save() *redis.StatusCmd {
 		return m.client.Save()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Shutdown() *redis.StatusCmd {
@@ -1660,9 +822,7 @@ func (m *ClientMock) Shutdown() *redis.StatusCmd {
 		return m.client.Shutdown()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ShutdownSave() *redis.StatusCmd {
@@ -1670,9 +830,7 @@ func (m *ClientMock) ShutdownSave() *redis.StatusCmd {
 		return m.client.ShutdownSave()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ShutdownNoSave() *redis.StatusCmd {
@@ -1680,9 +838,7 @@ func (m *ClientMock) ShutdownNoSave() *redis.StatusCmd {
 		return m.client.ShutdownNoSave()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) SlaveOf(host, port string) *redis.StatusCmd {
@@ -1690,9 +846,7 @@ func (m *ClientMock) SlaveOf(host, port string) *redis.StatusCmd {
 		return m.client.SlaveOf(host, port)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) Time() *redis.TimeCmd {
@@ -1700,9 +854,7 @@ func (m *ClientMock) Time() *redis.TimeCmd {
 		return m.client.Time()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.TimeCmd)
+	return m.Called().Get(0).(*redis.TimeCmd)
 }
 
 func (m *ClientMock) Eval(script string, keys []string, args ...interface{}) *redis.Cmd {
@@ -1710,9 +862,7 @@ func (m *ClientMock) Eval(script string, keys []string, args ...interface{}) *re
 		return m.client.Eval(script, keys, args...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.Cmd)
+	return m.Called().Get(0).(*redis.Cmd)
 }
 
 func (m *ClientMock) EvalSha(sha1 string, keys []string, args ...interface{}) *redis.Cmd {
@@ -1720,9 +870,7 @@ func (m *ClientMock) EvalSha(sha1 string, keys []string, args ...interface{}) *r
 		return m.client.EvalSha(sha1, keys, args...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.Cmd)
+	return m.Called().Get(0).(*redis.Cmd)
 }
 
 func (m *ClientMock) ScriptExists(hashes ...string) *redis.BoolSliceCmd {
@@ -1730,9 +878,7 @@ func (m *ClientMock) ScriptExists(hashes ...string) *redis.BoolSliceCmd {
 		return m.client.ScriptExists()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.BoolSliceCmd)
+	return m.Called().Get(0).(*redis.BoolSliceCmd)
 }
 
 func (m *ClientMock) ScriptFlush() *redis.StatusCmd {
@@ -1740,9 +886,7 @@ func (m *ClientMock) ScriptFlush() *redis.StatusCmd {
 		return m.client.ScriptFlush()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ScriptKill() *redis.StatusCmd {
@@ -1750,9 +894,7 @@ func (m *ClientMock) ScriptKill() *redis.StatusCmd {
 		return m.client.ScriptKill()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ScriptLoad(script string) *redis.StringCmd {
@@ -1760,9 +902,7 @@ func (m *ClientMock) ScriptLoad(script string) *redis.StringCmd {
 		return m.client.ScriptLoad(script)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) DebugObject(key string) *redis.StringCmd {
@@ -1770,9 +910,7 @@ func (m *ClientMock) DebugObject(key string) *redis.StringCmd {
 		return m.client.DebugObject(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) Publish(channel string, message interface{}) *redis.IntCmd {
@@ -1780,9 +918,7 @@ func (m *ClientMock) Publish(channel string, message interface{}) *redis.IntCmd 
 		return m.client.Publish(channel, message)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) PubSubChannels(pattern string) *redis.StringSliceCmd {
@@ -1790,9 +926,7 @@ func (m *ClientMock) PubSubChannels(pattern string) *redis.StringSliceCmd {
 		return m.client.PubSubChannels(pattern)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) PubSubNumSub(channels ...string) *redis.StringIntMapCmd {
@@ -1800,9 +934,7 @@ func (m *ClientMock) PubSubNumSub(channels ...string) *redis.StringIntMapCmd {
 		return m.client.PubSubNumSub()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringIntMapCmd)
+	return m.Called().Get(0).(*redis.StringIntMapCmd)
 }
 
 func (m *ClientMock) PubSubNumPat() *redis.IntCmd {
@@ -1810,9 +942,7 @@ func (m *ClientMock) PubSubNumPat() *redis.IntCmd {
 		return m.client.PubSubNumPat()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ClusterSlots() *redis.ClusterSlotsCmd {
@@ -1820,9 +950,7 @@ func (m *ClientMock) ClusterSlots() *redis.ClusterSlotsCmd {
 		return m.client.ClusterSlots()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.ClusterSlotsCmd)
+	return m.Called().Get(0).(*redis.ClusterSlotsCmd)
 }
 
 func (m *ClientMock) ClusterNodes() *redis.StringCmd {
@@ -1830,9 +958,7 @@ func (m *ClientMock) ClusterNodes() *redis.StringCmd {
 		return m.client.ClusterNodes()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) ClusterMeet(host, port string) *redis.StatusCmd {
@@ -1840,9 +966,7 @@ func (m *ClientMock) ClusterMeet(host, port string) *redis.StatusCmd {
 		return m.client.ClusterMeet(host, port)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterForget(nodeID string) *redis.StatusCmd {
@@ -1850,9 +974,7 @@ func (m *ClientMock) ClusterForget(nodeID string) *redis.StatusCmd {
 		return m.client.ClusterForget(nodeID)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterReplicate(nodeID string) *redis.StatusCmd {
@@ -1860,9 +982,7 @@ func (m *ClientMock) ClusterReplicate(nodeID string) *redis.StatusCmd {
 		return m.client.ClusterReplicate(nodeID)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterResetSoft() *redis.StatusCmd {
@@ -1870,9 +990,7 @@ func (m *ClientMock) ClusterResetSoft() *redis.StatusCmd {
 		return m.client.ClusterResetSoft()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterResetHard() *redis.StatusCmd {
@@ -1880,9 +998,7 @@ func (m *ClientMock) ClusterResetHard() *redis.StatusCmd {
 		return m.client.ClusterResetHard()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterInfo() *redis.StringCmd {
@@ -1890,9 +1006,7 @@ func (m *ClientMock) ClusterInfo() *redis.StringCmd {
 		return m.client.ClusterInfo()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
+	return m.Called().Get(0).(*redis.StringCmd)
 }
 
 func (m *ClientMock) ClusterKeySlot(key string) *redis.IntCmd {
@@ -1900,9 +1014,7 @@ func (m *ClientMock) ClusterKeySlot(key string) *redis.IntCmd {
 		return m.client.ClusterKeySlot(key)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ClusterCountFailureReports(nodeID string) *redis.IntCmd {
@@ -1910,9 +1022,7 @@ func (m *ClientMock) ClusterCountFailureReports(nodeID string) *redis.IntCmd {
 		return m.client.ClusterCountFailureReports(nodeID)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ClusterCountKeysInSlot(slot int) *redis.IntCmd {
@@ -1920,9 +1030,7 @@ func (m *ClientMock) ClusterCountKeysInSlot(slot int) *redis.IntCmd {
 		return m.client.ClusterCountKeysInSlot(slot)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ClusterDelSlots(slots ...int) *redis.StatusCmd {
@@ -1930,9 +1038,7 @@ func (m *ClientMock) ClusterDelSlots(slots ...int) *redis.StatusCmd {
 		return m.client.ClusterDelSlots(slots...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterDelSlotsRange(min, max int) *redis.StatusCmd {
@@ -1940,9 +1046,7 @@ func (m *ClientMock) ClusterDelSlotsRange(min, max int) *redis.StatusCmd {
 		return m.client.ClusterDelSlotsRange(min, max)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterSaveConfig() *redis.StatusCmd {
@@ -1950,9 +1054,7 @@ func (m *ClientMock) ClusterSaveConfig() *redis.StatusCmd {
 		return m.client.ClusterSaveConfig()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterSlaves(nodeID string) *redis.StringSliceCmd {
@@ -1960,9 +1062,7 @@ func (m *ClientMock) ClusterSlaves(nodeID string) *redis.StringSliceCmd {
 		return m.client.ClusterSlaves(nodeID)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) ClusterFailover() *redis.StatusCmd {
@@ -1970,9 +1070,7 @@ func (m *ClientMock) ClusterFailover() *redis.StatusCmd {
 		return m.client.ClusterFailover()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterAddSlots(slots ...int) *redis.StatusCmd {
@@ -1980,9 +1078,7 @@ func (m *ClientMock) ClusterAddSlots(slots ...int) *redis.StatusCmd {
 		return m.client.ClusterAddSlots()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ClusterAddSlotsRange(min, max int) *redis.StatusCmd {
@@ -1990,9 +1086,7 @@ func (m *ClientMock) ClusterAddSlotsRange(min, max int) *redis.StatusCmd {
 		return m.client.ClusterAddSlotsRange(min, max)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) GeoAdd(key string, geoLocation ...*redis.GeoLocation) *redis.IntCmd {
@@ -2000,9 +1094,7 @@ func (m *ClientMock) GeoAdd(key string, geoLocation ...*redis.GeoLocation) *redi
 		return m.client.GeoAdd(key, geoLocation...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) GeoPos(key string, members ...string) *redis.GeoPosCmd {
@@ -2010,9 +1102,7 @@ func (m *ClientMock) GeoPos(key string, members ...string) *redis.GeoPosCmd {
 		return m.client.GeoPos(key, members...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.GeoPosCmd)
+	return m.Called().Get(0).(*redis.GeoPosCmd)
 }
 
 func (m *ClientMock) GeoRadius(key string, longitude, latitude float64, query *redis.GeoRadiusQuery) *redis.GeoLocationCmd {
@@ -2020,9 +1110,7 @@ func (m *ClientMock) GeoRadius(key string, longitude, latitude float64, query *r
 		return m.client.GeoRadius(key, longitude, latitude, query)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.GeoLocationCmd)
+	return m.Called().Get(0).(*redis.GeoLocationCmd)
 }
 
 func (m *ClientMock) GeoRadiusRO(key string, longitude, latitude float64, query *redis.GeoRadiusQuery) *redis.GeoLocationCmd {
@@ -2030,9 +1118,7 @@ func (m *ClientMock) GeoRadiusRO(key string, longitude, latitude float64, query 
 		return m.client.GeoRadiusRO(key, longitude, latitude, query)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.GeoLocationCmd)
+	return m.Called().Get(0).(*redis.GeoLocationCmd)
 }
 
 func (m *ClientMock) GeoRadiusByMember(key, member string, query *redis.GeoRadiusQuery) *redis.GeoLocationCmd {
@@ -2040,9 +1126,7 @@ func (m *ClientMock) GeoRadiusByMember(key, member string, query *redis.GeoRadiu
 		return m.client.GeoRadiusByMember(key, member, query)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.GeoLocationCmd)
+	return m.Called().Get(0).(*redis.GeoLocationCmd)
 }
 
 func (m *ClientMock) GeoRadiusByMemberRO(key, member string, query *redis.GeoRadiusQuery) *redis.GeoLocationCmd {
@@ -2050,9 +1134,7 @@ func (m *ClientMock) GeoRadiusByMemberRO(key, member string, query *redis.GeoRad
 		return m.client.GeoRadiusByMemberRO(key, member, query)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.GeoLocationCmd)
+	return m.Called().Get(0).(*redis.GeoLocationCmd)
 }
 
 func (m *ClientMock) GeoDist(key string, member1, member2, unit string) *redis.FloatCmd {
@@ -2060,9 +1142,7 @@ func (m *ClientMock) GeoDist(key string, member1, member2, unit string) *redis.F
 		return m.client.GeoDist(key, member1, member2, unit)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.FloatCmd)
+	return m.Called().Get(0).(*redis.FloatCmd)
 }
 
 func (m *ClientMock) GeoHash(key string, members ...string) *redis.StringSliceCmd {
@@ -2070,9 +1150,7 @@ func (m *ClientMock) GeoHash(key string, members ...string) *redis.StringSliceCm
 		return m.client.GeoHash(key, members...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringSliceCmd)
+	return m.Called().Get(0).(*redis.StringSliceCmd)
 }
 
 func (m *ClientMock) Command() *redis.CommandsInfoCmd {
@@ -2080,9 +1158,7 @@ func (m *ClientMock) Command() *redis.CommandsInfoCmd {
 		return m.client.Command()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.CommandsInfoCmd)
+	return m.Called().Get(0).(*redis.CommandsInfoCmd)
 }
 
 func (m *ClientMock) ClientKillByFilter(keys ...string) *redis.IntCmd {
@@ -2090,9 +1166,7 @@ func (m *ClientMock) ClientKillByFilter(keys ...string) *redis.IntCmd {
 		return m.client.ClientKillByFilter(keys...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) MemoryUsage(key string, samples ...int) *redis.IntCmd {
@@ -2100,9 +1174,7 @@ func (m *ClientMock) MemoryUsage(key string, samples ...int) *redis.IntCmd {
 		return m.client.MemoryUsage(key, samples...)
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
+	return m.Called().Get(0).(*redis.IntCmd)
 }
 
 func (m *ClientMock) ReadOnly() *redis.StatusCmd {
@@ -2110,9 +1182,7 @@ func (m *ClientMock) ReadOnly() *redis.StatusCmd {
 		return m.client.ReadOnly()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
 
 func (m *ClientMock) ReadWrite() *redis.StatusCmd {
@@ -2120,107 +1190,5 @@ func (m *ClientMock) ReadWrite() *redis.StatusCmd {
 		return m.client.ReadWrite()
 	}
 
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StatusCmd)
-}
-
-func (m *ClientMock) XAdd(stream, id string, values map[string]interface{}) *redis.StringCmd {
-	if !m.hasStub("XAdd") {
-		return m.client.XAdd(stream, id, values)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
-}
-
-func (m *ClientMock) XLen(stream string) *redis.IntCmd {
-	if !m.hasStub("XLen") {
-		return m.client.XLen(stream)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.IntCmd)
-}
-
-func (m *ClientMock) XAddExt(opt *redis.XAddExt) *redis.StringCmd {
-	if !m.hasStub("XAddExt") {
-		return m.client.XAddExt(opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.StringCmd)
-}
-
-func (m *ClientMock) XRange(stream, start, stop string) *redis.XMessageSliceCmd {
-	if !m.hasStub("XRange") {
-		return m.client.XRange(stream, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XMessageSliceCmd)
-}
-
-func (m *ClientMock) XRangeN(stream, start, stop string, count int64) *redis.XMessageSliceCmd {
-	if !m.hasStub("XRangeN") {
-		return m.client.XRangeN(stream, start, stop, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XMessageSliceCmd)
-}
-
-func (m *ClientMock) XRead(streams ...string) *redis.XStreamSliceCmd {
-	if !m.hasStub("XRead") {
-		return m.client.XRead(streams...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XStreamSliceCmd)
-}
-
-func (m *ClientMock) XReadN(count int64, streams ...string) *redis.XStreamSliceCmd {
-	if !m.hasStub("XReadN") {
-		return m.client.XReadN(count, streams...)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XStreamSliceCmd)
-}
-
-func (m *ClientMock) XReadExt(opt *redis.XReadExt) *redis.XStreamSliceCmd {
-	if !m.hasStub("XReadExt") {
-		return m.client.XReadExt(opt)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XStreamSliceCmd)
-}
-
-func (m *ClientMock) XRevRange(stream string, start, stop string) *redis.XMessageSliceCmd {
-	if !m.hasStub("XRevRange") {
-		return m.client.XRevRange(stream, start, stop)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XMessageSliceCmd)
-}
-
-func (m *ClientMock) XRevRangeN(stream string, start, stop string, count int64) *redis.XMessageSliceCmd {
-	if !m.hasStub("XRevRangeN") {
-		return m.client.XRevRangeN(stream, start, stop, count)
-	}
-
-	mockArgs := m.Called()
-
-	return mockArgs.Get(0).(*redis.XMessageSliceCmd)
+	return m.Called().Get(0).(*redis.StatusCmd)
 }
